+++ src/momo_args.h
@@ -86,6 +86,8 @@ struct MomoArgs {
   std::string ayame_room_id;
   std::string ayame_client_id = "";
   std::string ayame_signaling_key = "";
+  // sendrecv, sendonly, recvonly
+  std::string ayame_direction = "sendrecv";
   // 空文字の場合コーデックは WebRTC デフォルトを使用
   std::string ayame_video_codec_type = "";
   std::string ayame_audio_codec_type = "";