+++ src/util.cpp
@@ -268,6 +268,9 @@ void Util::ParseArgs(int argc,
   ayame_app->add_option("--client-id", args.ayame_client_id, "Client ID");
   ayame_app->add_option("--signaling-key", args.ayame_signaling_key,
                         "Signaling key");
+  ayame_app->add_option("--direction", args.ayame_direction,
+                        "Direction (default: sendrecv)")
+      ->check(CLI::IsMember({"sendrecv", "sendonly", "recvonly"}));
   ayame_app
       ->add_option("--video-codec-type", args.ayame_video_codec_type,
                    "Video codec type (VP8, VP9, AV1, H264, H265)")